+++ src/modules/esm.mjs
@@ -1,0 +1,38 @@
+import path from 'path';
+import { release, version } from 'node:os';
+import { createServer as createServerHttp } from 'http';
+import './files/c.js';
+
+const random = Math.random();
+const __filename = new URL(import.meta.url).pathname;
+const __dirname = path.dirname(new URL(import.meta.url).pathname);
+
+let unknownObject;
+
+if (random > 0.5) {
+  unknownObject = path.resolve('./files/a.json');
+} else {
+  unknownObject = path.resolve('./files/b.json');
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const myServer = createServerHttp((_, res) => {
+  res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+  console.log(`Server is listening on port ${PORT}`);
+  console.log('To terminate it, use Ctrl+C combination');
+});
+
+export default { unknownObject, myServer };
